Update import statements after file renames
